// packages/core/src/odata-client/v4/edm-types.ts
/* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
export type EdmType =
  | 'Edm.Binary'
  | 'Edm.Boolean'
  | 'Edm.Byte'
  | 'Edm.Date'
  | 'Edm.DateTimeOffset'
  | 'Edm.Decimal'
  | 'Edm.Double'
  | 'Edm.Duration'
  | 'Edm.Float'
  | 'Edm.Guid'
  | 'Edm.Int16'
  | 'Edm.Int32'
  | 'Edm.Int64'
  | 'Edm.SByte'
  | 'Edm.Single'
  | 'Edm.Stream'
  | 'Edm.String'
  | 'Edm.TimeOfDay';
// + Geography
// + Geometry